# packages/sdk/openapi/agentflow.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentFlow API",
    "version": "0.1.0",
    "description": "REST API for AgentFlow — an open-source workflow automation platform.\n\n**Authentication** — send `Authorization: Bearer <token>` with a JWT access token (POST /api/auth/login) or a personal API key (`af_…`, POST /api/api-keys).\n\n**Rate limiting** — 200 requests/minute per user or IP by default; stricter limits apply to auth, AI and checkout endpoints.\n\n**Pagination** — list endpoints accept `page` and `limit` (max 100) and return the applied values in the `X-Page` and `X-Limit` response headers.",
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/api/credentials": {
      "get": {
        "tags": [
          "Credentials"
        ],
        "summary": "List credentials (masked)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            },
            "required": true,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "example": 25
            },
            "required": true,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Credentials without secret values",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Credential"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Credentials"
        ],
        "summary": "Store an encrypted credential",
        "description": "Values are encrypted at rest with AES-256-GCM (CREDENTIAL_ENCRYPTION_KEY).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Credential to store",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "type": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "api_key",
                          "bearer_token",
                          "basic_auth",
                          "oauth2_managed",
                          "oauth2_custom",
                          "header_auth",
                          "query_auth",
                          "mcp_oauth2",
                          "oauth2",
                          "basic",
                          "token"
                        ]
                      },
                      {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 50
                      }
                    ]
                  },
                  "provider": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "name",
                  "type",
                  "provider",
                  "data"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stored credential (masked)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/credentials/{id}/reveal": {
      "get": {
        "tags": [
          "Credentials"
        ],
        "summary": "Reveal a credential's decrypted value",
        "description": "Restricted to organization OWNER and ADMIN members.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Entity ID (cuid)"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Decrypted credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialReveal"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Only owners and admins can reveal credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Credential not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/credentials/{id}": {
      "delete": {
        "tags": [
          "Credentials"
        ],
        "summary": "Delete a credential",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Entity ID (cuid)"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Credential not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code",
            "example": "NOT_FOUND"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Field-level validation errors (VALIDATION_ERROR only)"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "description": "Error envelope used by every non-2xx response"
      },
...
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "Credential": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Slack prod token"
          },
          "type": {
            "type": "string",
            "enum": [
              "api_key",
              "oauth2",
              "basic",
              "token"
            ]
          },
          "provider": {
            "type": "string",
            "example": "slack"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "data": {
            "type": "object",
            "properties": {
              "hasValue": {
                "type": "boolean",
                "enum": [
                  true
                ]
              }
            },
            "required": [
              "hasValue"
            ],
            "description": "Value is never returned in listings"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "provider",
          "createdAt",
          "data"
        ]
      },
      "CredentialReveal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "api_key",
              "oauth2",
              "basic",
              "token"
            ]
          },
          "provider": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": {},
            "description": "Decrypted credential payload (OWNER/ADMIN only)"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "provider",
          "data"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "AgentFlow access token (from POST /api/auth/login) or a personal API key (`af_…`, from POST /api/api-keys). Send as `Authorization: Bearer <token>`."
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Signature",
        "description": "HMAC-SHA256 hex digest of the raw request body, computed with the webhook secret (optional `sha256=` prefix). Required on every /api/webhooks/trigger/{path} call."
      },
      "stripeSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature",
        "description": "Stripe webhook signature header (t=…,v1=…)."
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    },
    {
      "url": "https://api.agentflow.dev",
      "description": "Production (placeholder)"
    }
  ],
...
}